+++ packages/article-paragraph/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.21](https://github.com/newsuk/times-components/compare/@times-components/article-paragraph@1.9.20...@times-components/article-paragraph@1.9.21) (2023-06-01)
+
+**Note:** Version bump only for package @times-components/article-paragraph
+
+
+
+
+
 ## [1.9.20](https://github.com/newsuk/times-components/compare/@times-components/article-paragraph@1.9.19...@times-components/article-paragraph@1.9.20) (2023-05-15)
 
 **Note:** Version bump only for package @times-components/article-paragraph